refactor: Moves current learner files to the auth package.
