+++ src/app/infrastructure/auth/login/login.component.html
@@ -1,8 +1,8 @@
-<form [formGroup]="loginForm" (ngSubmit)="onLogin()" fxLayout="column">
-  <mat-error *ngIf="hasError">Indeks nije pronađen. Molimo vas prvo kreirajte nalog.</mat-error>
+<form [formGroup]="loginForm" (ngSubmit)="login()" fxLayout="column">
+  <mat-error *ngIf="hasError">Korisničko ime ili lozinka nisu ispravni.</mat-error>
   <mat-form-field>
-    <mat-label>Indeks</mat-label>
-    <input matInput type="text" name="index" formControlName="username">
+    <mat-label>Korisničko ime</mat-label>
+    <input matInput type="text" name="username" formControlName="username">
   </mat-form-field>
   <mat-form-field>
     <mat-label>Lozinka</mat-label>
