+++ src/app/infrastructure/auth/login/login.component.ts
@@ -1,16 +1,16 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { LearnerService } from '../learner.service';
 import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
-import { InterfacingInstructor } from '../../instructor/interfacing-instructor.service';
+import { InterfacingInstructor } from '../../../modules/instructor/interfacing-instructor.service';
 
 @Component({
   selector: 'cc-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })
-export class LoginComponent implements OnInit {
+export class LoginComponent {
 
   loginForm = new FormGroup({
     username: new FormControl('', [ Validators.required ]),
@@ -23,10 +23,7 @@ export class LoginComponent implements OnInit {
     private router: Router,
     private instructor: InterfacingInstructor) { }
 
-  ngOnInit(): void {
-  }
-
-  onLogin(): void {
+  login(): void {
     if (this.loginForm.valid) {
       this.learnerService.login(this.loginForm.value).subscribe(() => {
         this.router.navigate(['/'])
