--- src/app/modules/domain/feedback/feedback.component.css
@@ -1,8 +1,0 @@
-.mat-button-toggle-group {
-  height: 32px;
-  align-items: center;
-}
-
-.center {
-  align-items: center;
-}
