+++ src/app/modules/domain/learning-objects/arrange-task/arrange-task.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
 import {environment} from '../../../../../environments/environment';
-import {LearnerService} from '../../../learner/learner.service';
+import {LearnerService} from '../../../../infrastructure/auth/learner.service';
 import {ArrangeTaskContainerSubmission} from './model/arrange-task-container-submission.model';
 import {map} from 'rxjs/operators';
 import {ArrangeTaskEvaluation} from './model/arrange-task-evaluation.model';
