+++ src/app/modules/domain/learning-objects/learning-object-container/learning-object-container.component.ts
@@ -2,7 +2,6 @@ import { Component, ComponentFactoryResolver, Input, OnInit, ViewChild } from '@
 import { LearningObjectDirective } from '../learning-object.directive';
 import { LearningObject } from '../learning-object.model';
 import { LearningObjectComponent } from '../learning-object-component';
-import {FeedbackService} from '../../feedback/feedback.service';
 
 @Component({
   selector: 'cc-learning-object-container',
@@ -14,15 +13,11 @@ export class LearningObjectContainerComponent implements OnInit {
   @Input() learningObject: LearningObject;
   @ViewChild(LearningObjectDirective, {static: true}) learningObjectHost: LearningObjectDirective;
 
-  constructor(private componentFactoryResolver: ComponentFactoryResolver, private feedbackService: FeedbackService) { }
+  constructor(private componentFactoryResolver: ComponentFactoryResolver) { }
 
   ngOnInit(): void {
     const componentFactory = this.componentFactoryResolver.resolveComponentFactory(this.learningObject.getComponent());
     const componentRef = this.learningObjectHost.viewContainerRef.createComponent<LearningObjectComponent>(componentFactory);
     componentRef.instance.learningObject = this.learningObject;
-  }
-
-  submitRating(rating: number): void {
-    this.feedbackService.submitRating(rating, this.learningObject.id);
   }
 }
