+++ src/app/modules/domain/learning-objects/short-answer-question/short-answer-question.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 import {HttpClient} from '@angular/common/http';
 import {environment} from '../../../../../environments/environment';
-import {LearnerService} from '../../../learner/learner.service';
+import {LearnerService} from '../../../../infrastructure/auth/learner.service';
 import {map} from 'rxjs/operators';
 import { SaqEvaluation } from './saq-evaluation.model';
