+++ src/app/modules/domain/notes/markdown-editor/markdown-editor.component.ts
@@ -1,6 +1,5 @@
 import {
   Component,
-  Input,
   ViewChild,
   forwardRef,
   ElementRef,
